repair coin index
temporary PR.
This PR introduces a background task to repair the `coin_index` and
remove any dangling entries. The PR will be active for one release and
will be reverted afterward.

The background task works by iterating over a snapshot of the
`coin_index`, identifying coins that no longer belong to their
respective owners, and populating a list of candidates for removal(some
entries might be benign)

Once the candidate list is populated, the task makes a second pass over
the candidates list. This time it locks the corresponding entries to
prevent potential races with concurrent writes. The task then reverifies
the eligibility criteria and removes the dangling entries