[CI] do not ignore failures from Split Cluster Check in PRs
## Description 

Currently when the check runs in PR CI, the failure is ignored and
bisect does not run. So the workflow succeeds when it shouldn't.
For example,
https://github.com/GeunhwaJeong/haneul/actions/runs/10641580676/job/29502904956#step:3:2268

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: