[consensus] Switch to certified vote scoring & incrementally score subdags
After analyzing the results of our Leader Scoring Strategy
experimentation. We will be switching from a Vote scoring strategy to a
Certified Vote Scoring Strategy. This has shown improvements of about
~90ms for P50 latency with 6K mixed TPS in private-testnet. We can see
that this scoring strategy gives us the best distribution of scores, as
scores are distributed together across major geographic regions which
improves our quorum latency.

As part of this change we also moved to an incremental scoring process
with a new struct called ScoringSubdag that keeps track of the votes for
leaders and relevant stake for eventual reputation score calculation.

Also removed the ReputationScoreCalculator & LeaderScoringStrategy
components as we are now finalizing on the scoring strategy we will be
using.

Full experiment results :
https://www.notion.so/haneullabs/Leader-Scoring-Strategy-f11bbbd1055e453f9f0f5490544941ed?pvs=4

## Testing
- [x] unit tests & simtests
- [ ] run certified vote v2 & v3 with incremental scoring in
private-testnet and finalize on one
- [ ] run upgrade test from vote scoring to certified vote scoring