[move][test_scenario] Fix bug in `test_scenario` where we weren't returning allocated tickets
## Description 

Fixes a bug where we weren't returning allocated receiving tickets in a
transaction at the end of the transaction. This fixes the bug, and adds
tests to make sure allocated tickets are properly handled at the end of
a test scenario transaction.

## Test plan 

Added new tests to test the failing behavior and made sure they pass.