[move-ide] Split symbolication implementation into smaller functions
## Description 

This PR contains a refactoring to split a super-long `get_symbols`
function into two smaller ones.

It not only makes code mode readable but also enables a more optimized
implementation of the test suite. Previously, all symbols had to be
recomputed when tests involved a cursor, even though it was only cursor
that was changing, rather than symbols themselves. With this PR, we can
recompute the cursor only, which brings the time to run all
`move-analyzer` tests from minutes to (tens of) seconds.

## Test plan 

All existing tests must pass