[trace-view] Added support for continue
## Description 

This PR adds support for "continue" debugger action. It also makes
implementation of "next" ("step over") and "step out" actions faithful
to their intended semantics (i.e., executing all instruction in a
function stepped over or stepped out of). These changes do not actually
change the current behavior of the viewer - they are done in preparation
for implementing variable value tracking and displaying.

This PR also includes a refinement of the "step out" action
implementation. As also explained in the code comment, previously in the
following line of code, after entering `foo` and stepping out of it, we
would immediately step into `bar`, which could be quite confusing for
the user:
```
assert(foo() == bar());
```

Finally, this PR also includes some formatting changes for lines that
were a bit too long

## Test plan 

Tested manually