[haneul_proxy/pod health]
## Description
* add a route for pod health in k8s

i am using the consumer_operations_submitted value but more values can
be added later if we need to. when this value is non-zero, it means the
service is processing data correctly and can be used to consider it
healthy enough to handle traffic. it will be implemented as a liveness
check.

## Test Plan
local

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: