[cli] Remove indexer feature for `haneul` and remove the `haneul-pg` binary
## Description 

Recent improvements on GraphQL and Postgres code removed the dynamic
linking to `libpq`. Due to this linking (GH actions was linking to
postgres@14), we had to have a separate `haneul-pg` binary that was built
with the `indexer` feature such that one can have access to `haneul start
--with-indexer` and `--with-graphql` due to the required dependencies to
Postgres.

This PR removes the `indexer` feature, all the instances of building the
binary with `--features indexer` in workflows, actions, etc, and updates
the documentation (docs + haneul-test-validator crate). This is possible
because `libpq` is no longer a dependency that gets dynamically linked
to the binary. Note that in order to use those flags, a running Postgres
DB is still required just like before.

## Test plan 

Existing tests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: the `indexer` feature was removed from the `haneul` crate as the
dynamic linking to `libpq` was removed. Therefore, the `haneul-pg` binary
will not be part of releases anymore. This `haneul-pg` binary was used for
starting a network with `--with-indexer` and `--with-graphql` flags.
These commands will still work as before and it is still required to
have installed a Postgres database.
If you used `haneul-pg` binary previously, you can simply use `haneul` binary
from this version on.
- [ ] Rust SDK:
- [ ] REST API: