[move lockfile] rename dependency `name` field to `id` and add a separate `name` field to store manifest name
## Description

As discussed in
https://github.com/GeunhwaJeong/haneul/pull/14178#issuecomment-1924760426
(second and third point), a few changes to the lockfile:
- the `name` field in `dependencies` is renamed to `id` to better
reflect the meaning in the dependency graph (the packages are
discriminated by their identifier, as resolved by the hook, which is not
necessarily their manifest name)
- added a `name` field which will store the dependency manifest name
(this is needed to show user-friendly error messages using the package
manifest name instead of identifier which may be confusing)
- bumped lockfile version to 3

This is part of the work to enable compiling against on-chain
dependencies https://github.com/GeunhwaJeong/haneul/pull/14178.

cc rvantonder amnn

## Test plan 

All changes are covered by existing unit tests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: Introduce lock file version 3, which renames a dependency's
`name` field to `id` and introduces a separate `name` field that stores
the package's name from its manifest for improved error reporting. Older
lock files will be ignored when resolving dependencies so that this
information is always available.
- [ ] Rust SDK:
- [ ] REST API: