[bridge-indexer] move current checkpoint metric to retrieval stage
## Description 

Currently indexer reports `current_checkpoint` when saving progress to
DB. This makes the metrics a bit out of state when we cache the metrics
in memory. This PR:
1. moves the metrics to data retrieval stage,
2. rename it to `latest_retrieved_checkpoints`
3. in EthSubscription task, periodically fetch the latest block height
and update this metric

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: