[GraphQL] Rework Event type
## Description 

Accessing the `bcs` field of an `Event` is rather confusing, as it is
not the whole event BCS encoded, but just the contents of it, more
specifically, the BCS bytes of a `MoveValue`.

This PR removes the #flatten in `Event`, adds a `contents` field to the
type, and adds the `bcs` field to the `Event` type which encodes the
whole `Event`.

## Test plan 

Existing tests

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [x] GraphQL: The `Event` type has a new field called `contents`, which
is the event's content value as a `MoveValue`. This replaces the
previous scheme that flattened the `MoveValue` type in the `Event` type.
A `bcs` field was also added, which represents the Base64 encoded BCS
serialized event.
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: