[AuthorityAggregator] EpochEnded to be retryable
## Description 

E2e localnet tests appear to fail sporadically due the the `EpochEnded`
error. Although it should not be the common case it's possible to happen
when a transaction has reached a node's RPC endpoint, but epoch changes
in the middle. In this case an `EpochEnded` can be thrown when trying to
access the AuthorityEpochStore. Making the error retryable will allow
the quorum driver retry the transaction and hopefully succeed.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: