graphql: ObjectOwner::Parent exposed as Owner
## Description

Now that we no longer expose wrapped objects, we need a way to expose
the addresses of object parents when they are other objects that have
been wrapped.

## Test plan

Updated tests:

```
haneul$ cargo nextest run -p haneul-graphql-e2e-tests
haneul$ cargo nextest run -p haneul-graphql-rpc
```

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [x] GraphQL: Change `Parent.parent` from an `Object` to an `Owner`.
Although it's guaranteed to be an object if it exists, it may be
wrapped, in which case it will not exist. Exposing it as an Owner allows
queries to extract its ID and also fetch other dynamic fields from it
even if it is wrapped.
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: