[Consensus] Disable periodic synchronizer when commit is lagging and far missing blocks
## Description 

This is fixing/improving the behaviour of the periodic synchronizer to:
(1) allow the backfilling of missing blocks when commit is lagging but
we do have missing blocks that are in an tolerable threshold range from
the current highest accepted round
(2) stop the periodic synchronizer when commit is lagging and missing
blocks are too far in the future

Point (1) will allow bypassing a recently observed issue where
equivocating blocks make their way into the DAG but they are not
committed (due to the current commit rule this is not allowed). Then
nodes that are using the commit syncer to catch up might come across
blocks that appear on the fetched committed sub dags but do have
ancestor dependencies on equivocated blocks that never got committed.
Currently as the synchronizer is disabled nothing will attempts to fetch
the missing block effectively making the catch up stop. Now with change
(1) will be able to do so.

## Test plan 

CI/PT

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: