graphql: expose MovePackage as BCS (packageBcs)
## Description

Add a new field -- `MovePackage.packageBcs` -- that exposes the BCS
representation of the MovePackage. This complements `bcs` (which is the
BCS representation of its `Object` form), and `moduleBcs` (which is the
BCS representation of its `Vec<CompiledModule>`s).

## Test plan

New E2E test:

```
haneul$ cargo nextest run -p haneul-graphql-e2e-tests -- packages/bcs.move
```

## Stack

-  

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [x] GraphQL: Adds `MovePackage.packageBcs` to expose the BCS
representation of the `MovePackage` struct (as opposed to the BCS
representation of the outer `Object` struct.
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: