[move][docgen] Render string constants for error constants
## Description 

This updates how constants annotated with `#[error]` are rendered in
docgen. This also adds the `#[error]` annotation on them in the
generated documentation.

Note that we don't try to render all bytearrays as strings, but only
constants with the `#[error]` annotation to avoid rendering normal
bytearrays strings.

## Test plan 

Added a test to make sure we render error-annotated consts as we expect.