[indexer][writer] Add first_tx_sequence_number to epochs table to decouple from checkpoints table
## Description 

The `epoch_total_transactions` field on `epochs` table is calculated
today from `checkpoints.network_total_transactions`, which is not ideal
since the latter is prunable and `epochs` will not be pruned for the
foreseeable future. To remove this dependency, we add
`first_tx_sequence_number` to the `epochs` table at epoch boundary. That
is, the network total transaction count from the final checkpoint of the
epoch becomes the first tx sequence number of the new epoch. This also
means that at epoch boundary, the current-to-be-previous epoch's
`epoch_total_transactions` is derived from the checkpoint's network
total transactions - the epoch's `first_tx_sequence_number`.
Consequently, this will also help in the pruner implementation, as given
an epoch we'd like to know the corresponding cp and tx.

This encompasses just the writer change. Before updating the read path,
we will need to backfill the instance.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: