indexer-alt: support checkpoint upperbound for graceful shutdown
## Description
Previously, the service would just hang after having wound down all
pipelines during a graceful shutdown. This is because the graceful
shutdown process was left waiting for an explicit interrupt signal to
unblock the whole shutdown process.

This change relaxes that constraint: If the service is already shutting
down, then there is no need to wait for the explicit Ctrl-C. This makes
it feasible to run the indexer for specific ranges of checkpoints.

Note that we don't want to call the global cancellation token once
ingestion is done, because that might stop a downstream task before it
has acted on the downloaded checkpoints. Instead, the ingestion service
communicates completion by closing its channel, which propagates through
the handlers and committers.

## Test plan

```
RUST_LOG="info,haneul_indexer_alt=debug" cargo run -p haneul-indexer-alt --release --  \
  --database-url "postgres://postgres:postgrespw@localhost:5432/haneul_indexer_alt" \
  --remote-store-url https://checkpoints.mainnet.haneul.io                          \
  --last-checkpoint 1000
```