fix(indexer-alt): Flush pending rows before shutdown
## Description

When running the indexer on a finite range of checkpoints. Make sure
commiters' buffers are completely empty before shutting down their task,
otherwise we may not fully write out all the intended rows for the range
of checkpoints provided (there may be some data left at the bottom of
the barrel).

## Test plan

Ran the following:

```
cargo run -p haneul-indexer-alt --release -- \
  --database-url "postgres://postgres:postgrespw@localhost:5432/haneul_indexer_alt" \
  --remote-store-url https://checkpoints.mainnet.haneul.io \
  --last-checkpoint 2000
```

Corroborated that the data that results in the DB at the end:

- Stops at the expected checkpoint (not before or after)
- Matches counts of rows in the production mainnet DB for the equivalent
  tables at the same checkpoints.

This can/should be made into an automated test, but that requires tempdb
and migrations to be implemented (a comment has been added to this
effect).