[Indexer] Add synthetic ingestion and benchmark framework
## Description 

This PR adds two things:
1. A synthetic workload generator for the indexer. It generates
synthetic checkpoints and transactions which are written to the
ingestion dir. Currently only transfer transactions are generated, but
we could extend it later.
2. A benchmark framework that allows us to connect the synthetic
ingestion with any indexer implementation. The indexer implementation
will need to implement a wrapper for a trait type.

## Test plan 

Added an e2e test.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: