[Bridge] Retry on finalized transaction not observed
## Description 

In some cases, the client observes a finalized transaction before most
bridge authorities. In such cases, today the code will return an error
without retries, causing this validator to be skipped in terms of
signature aggregation. If bridge validators are using slow ethereum
fullnode providers, for example, this can be a large amount of the
committee, resulting in failed signature aggregation attempts in some
cases (due to not achieving quorum), or high gas costs in others (due to
not getting fewer higher staked validator signatures).

The solution here is to add retry logic in the map function.

## Test plan 

Will run on a testnet node and see what happens.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: