Uncompressed G1 group elements
## Description 

Currently, group elements are represented in compressed form and must be
decompressed before each operation. This is a rather expensive operation
to do, in particular if many elements are used.

This PR introduces a new uncompressed representation of BLS12381-G1
elements and a function to add a vector of uncompressed elements. For
500 terms, computing a sum will be about 100x faster to compute than
using compressed representation.

## Test plan 

Unit tests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [x] Protocol: Introduce an uncompressed representation of BLS12381-G1
group elements for faster addition of group elements.
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API:

---------